demote primitives to free variables in a bigger env
Thereupon it was discovered issues: Types in the bigger environment
are supposed to be binding. Current version's main method should
throw "Can't unify fixed name β = ω". The error is due to the name
"β" clashing with type variables generated in GenerativeNameGenerator.

In the next version, we shall fix the issue by having
GenerativeNameGenerator generating IDs.