let inferFrom take partial functions
(and rename things to my liking)

Example usage: Γ in TestUnification.main. It is an environment that
implements typing of integer literals, of which there are an infinite
number.

If global environments can be partial functions everywhere, then
we are free to leave all language primitives as free names, and
handle them modularly in the global environment.