make type variables truly impredicative
Take care of what appears to be the last remaining case:
The false type in the operator position. There should
be no other case, for the following reason.

There are 4 type constructors:
-  ∀ (quantification)
-  ★ (type application)
-  → (function type)
-  α (type variable)

Case ∀ is stripped away by a call to peelAwayQuantifiers.

Case ★ is always illegal, because it's in head normal form
(e. g., List α)

Case → is expected.

Case α is possible, and pertains two possibilities.

1. Operator type = ∀α. α, we handle it in this commit.

2. Operator type = α for some free name α. Since all free type
   names are in the environment, and the environment never
   provides a function type, it is also illegal.