update paper according to knowledge of Griffin's notion
Timothy G. Griffin.
A formula-as-type notion of control.

Changes are 2.

1. Realize that what I used to call "call/cc" is not the real
call/cc of scheme. If the argument of my previous call/cc does
not call the given continuation, then the current evaluation
context is discarded. My previous call/cc's behavior is actually
that of the operator C of idealized scheme. I call it "abort/cc"
for "abort with current continuation" from now on, because it
aborts to whatever its argument cares to provide, passing its
argument the current continuation on the way.

2. Discard Kolmogorov's embedding, use Fischer's embedding.
The cps tranformation becomes very much simpler. I have to
think of an embedding for universally quantified types,
though. Will have to look up whether that's new.