remove circularity test (not necessary anymore), discover flaw
Flaw: we need a rule about type abstraction.

We have assumed that whenever a lambda abstraction is used
polymorphically, there will be an argument type annotation
conveying the needed polymorphism. This is not true in the
body of matchNat:

  ns const (\abs x ys z f. f x (ys nil cons))

The last argument of `ns` is used polymorphpically without
there being a type annotation anywhere.

We must add CT-Tabs and prove things about it.